Fixed logic error in generating link desktop files
When appending a new URL, there was a check if the URL being added
was already in the list of localized URLs.  However, instead of
checking the values of the dictionary, we were checking the keys.
So, effectively it was always being added to the dictionary,
so no real harm was done, other than extra localizations were added
when they were not really necessary.  However, we could not simply
check the values with the existing logic, as that would have broken
the case where two locales have the same URL but different from
the default locale.  The correct logic is to check each URL against
that of the default locale.

[endlessm/eos-shell#2018]