Disable splash screen for billard-gl
The application does not properly notify the shell when
the window has loaded, and thus the splash screen remains
until the use closes it.  The application loads fast enough
that a splash screen is not that critical, so at least for now
let's simply disable the splash screen rather than patch the app.

[endlessm/eos-shell#2260]