Revert "Split icons for core vs. bundles"
This reverts commit 155203277a3339f0a8368ea614238e267c690f62.

Since the icons are used by the app store when a user hovers
over a thumbnail, for now let's keep all the icons shipped
as part of the core with eos-shell-content.
There will thus be duplication when the icon is installed
with the bundle, and at that time the icon inside the bundle
should pre-empt the one installed in the theme as part of the core.

Intentionally keeping the previous commit in the git history,
in case we ever reverse this decision.