Update default desktop config to match Global
Although we may separate them down the road,
for now the default config should be the same as global.
If there are apps that are not installed, they simply
will not show up on the desktop.

For now, let's keep them as separate files in case we
want to differentiate these desktop configs down the road,
but until that time whatever changes we want to make to Global
we should make to default as well.