Remove unsupported apps from Brazil icon grid layout
Even though listing these apps does not impact the view
of the desktop, if the apps are later available via the app store,
they would install into the specified folder rather than
directly on the desktop.

Also, don't put GoldenDict in a desktop folder, because it is not
properly translated to Portuguese, and will be soon replaced
with a link to wiktionary.org.

[endlessm/eos-shell#3622]