Check for any app with more than 3 screenshots
We only cleanly support up to 3 screenshots.
Issue a warning log to the terminal in this case.
Not a critical failure, but it can impact alignment
of the add details page of the app store at some resolutions.

Test the target rather than the source,
because the source can be .png or .jpg,
whereas the target always has a .jpg extension.
The CMS guarantees screenshot numerical sequencing,
so we can simply check for *4.jpg as an indicator
that there are at least 4 screenshots for the app.

[endlessm/eos-shell#3482]