Map deprecated app IDs
The unzip_content.py script reads renamed-apps.txt to get the mapping
from old app ID to new app ID.  This is reflected in the content.json
as a duplicate entry using the name and assets for the new app ID
but with special text for the subtitle and description that identifies
this old app ID as deprecated.  The old entry is inserted right before
the new entry, to help cluster the two entries together in the app store.

For now, this is only relevant for a few applicaitons that have been
delivered on the Guatemala images, so no concern that other translations
are missing -- we will pick them up on the next milestone.

[endlessm/eos-shell#4588]