Updates from CMS
Includes updated screenshot for English travel app.

Includes fix of Spanish link to pinterest.
The previous link was to an unofficial blog.
The official site is already properly translated to Spanish,
so let's use it instead.

Also, removed separate link id for English versions of
pinterest and tumblr, so that Spanish and English can
share the same assets.

[endlessm/eos-shell#4450]
[endlessm/eos-shell#4650]