Include English apps in Arabic icon grid
At the request of our investor in the middle east,
we are including English apps on the Arabic image
to demo what our full version looks like.
Let's do the same for the default icon grid layout,
which is now identical to the Global English version
except for the Arabic version of the encyclopedia.

[endlessm/eos-shell#4706]