tools: Add debhelper script for merging content
To merge the desktop files during package builds, add a debhelper
script, dh_eoscontent, that calls eos-content-merge. dh_eoscontent will
check that the app ID is in content.json. If it is and it finds a
desktop file in the install that matches the app ID, it will call
eos-content-merge.

Currently, dh_eoscontent needs to be manually invoked, but subsequent
commits will add integration to debhelper and CDBS.