Move converted core app desktop files to bundle directory
Eventually we want the core app desktop files to not be installed in
/usr/share/applications with Endless customizations merged into the
package's desktop file. Until the full conversion happens, we need to
keep the customized desktop files for packages that have been converted
out of /usr/share/appications so we don't override them at runtime.

Keep a hacky list of converted apps. If the app ID is in the list, the
desktop file will be installed into the bundles directory, which can be
used at build time.