eos-content-merge: Sort keys in rewritten desktop file
ConfigParser uses OrderedDict for each section, which maintains the
order that keys were added to it. That means that newly added keys will
always show up at the end of the "Desktop Entry" section.

The best we can do is create a new configuration, add the Desktop Entry
section with keys sorted, then merge in the rest of the rest of the
sections from the original. The sorting is taken directly from the
OrderedDict documentation.

[endlessm/eos-shell#4464]