eos-content-merge: Sort keys in rewritten desktop file
ConfigParser uses OrderedDict for each section, which maintains the
order that keys were added to it. That means that newly added keys will
always show up at the end of the "Desktop Entry" section.

For most fields this is the desired behavior, but the translated
"Name[..]" keys should be sorted alphabetically and immediately
follow the original "Name" key.  Though this has no impact to the
user experience, it helps when manually reviewing desktop files.

[endlessm/eos-shell#5336]