Remove empathy (chat) app
There are several fundamental problems with empathy.
It is pretty much a dead project upstream, the UI is quite bad,
and desktop chat integration with Endless is not very effective.
So, we have decided to stop supporting it.
In the future, we will completely remove it from the system,
but for now we at least remove it from the default desktop
configurations and from the app store (removed from the CMS).

[endlessm/eos-shell#5938]