Script to list installed/available apps
Reports details for a given software image, using files generated
by the image builder that can be downloaded from images.endlessm-sf.com,
and the metadata and translations that are part of eos-shell-content.

Includes core applications (those that are delivered as part of the OSTree),
installed app bundles, and app bundles that are available on the
app server but not installed by the image builder.

The available app bundles need further filtering to ignore localized
apps that are not made available in the app store based on locale.
Also, we should clean up the processing of available apps to specify
the latest app version that is available for the specified OS version.

It would also be useful to display the installed size of each app bundle
(not applicable to core apps).  This would be trivial for available apps,
but for installed apps we need to correlate with the size data that
we read from the app server, since these details are not available
on the image builder reports.

[endlessm/eos-shell#5911]