Add size calculation
Based on the size reported by the app server
(whether or not the app is pre-installed).
We report in MB and round to the nearest tenth.
If the server reports a size of 0 (e.g., for apps
posted to server prior to implementing size calculations),
we report the size as "Unknown".  For core apps,
the app is part of the OS itself and thus does not have
its own size, and we report the size as "System App".
Note that in the case of apps that download third party
content upon installation, we report here only the size
of the bundle on our server.

[endlessm/eos-shell#5911]