Don't use absolute paths for custom splash screens
We want to move to a place where splash screens are part of the bundle.
Before we do that, we also want to be free to move them around, while
they're part of the core OS.
Make sure the desktop files do not reference an absolute path to make
this possible.

[endlessm/eos-shell#5976]