data: Allow blacklisting of icon grid entries by arch
Some icon grid entries may not be appropriate on specific architectures,
so they need to be stripped from the final json files. To handle this, a
separate arch-blacklist.json is provided which can list entries that
should be blacklisted. A script, write-grid-json.py, is added to parse
the lists and remove the relevant entries.

The blacklist initially removes spotify on ARM since this requires
flash.

[endlessm/eos-shell#6012]

Conflicts:
	data/Makefile.am