write-grid-json: Write to temp file and rename
If there was an error dumping the json into the output file, it would
still exist truncated. Although make would catch the error from the
script, a subsequent run would see the existing target file and skip
running the recipe again.

Write to a temporary version of the output file and rename it to the
real output file name after successfully creating the temporary version.

[endlessm/eos-shell#6012]