Fix app reports for renamed core app IDs
Starting with OS release 2.6.0, we switched to new upstream
desktop names for gedit, nautilus (was eos-file-manager), and totem.
These no longer are exact matches to the package name, so we need
a special mapping for these.  For older OS releases, though,
we want to refer to the app ID as actually installed.

https://phabricator.endlessm.com/T11292