Reprocess all assets with latest imagemagick
With the combination of moving from 32-bit to 64-bit and/or
an upgrade to the imagemagick package, there is a slight binary
difference (with no user-visible change) on the generated assets.

Let's take the one-time hit now to make it easier to manage changes
moving forward, but let's not cherry-pick this for the eos2.6
stable branch (so as not to hit existing users with needless updates).
For eos3.0 we are switching to flatpak, so the assets would be
used differently regardless of this change.

https://phabricator.endlessm.com/T12047