Remove deprecated apps
These were removed a long time ago, but we kept the assets around
"just in case".  At this point, it is highly unlikely that we will
add any of these back (prefering third-party flatpak repos if these
apps are ever to be supported), so we removed them from the CMS
to avoid confusion.