Updates from CMS
Includes new Spotify app, update of KALite icon,
and rename of app ID for Skype.

Note that we still have the Spotify weblink for now,
but once the flatpak is published, we may want to
remove the weblink.

Also, don't localize KALite app IDs, as the localized content
will be delivered outside of the flatpak.