Unzip_content: round icons
Upon ingest from the CMS, all icons need to be modified
to apply the new corner rounding (15 pixel radius on a 60x60 square).
We will keep assets in the CMS with the old corner style.

Note that we now generate the image mask (with rounding)
on the fly using ImageMagick, so no need for the old mask file.

https://phabricator.endlessm.com/T13208