Updates from CMS
Includes weblinks for id, th, and vi.

Note that there is a one-time hit due to a change in ImageMagick
(or an underlying library) where previously any JPEG image that
did not have a density specified in pixels per inch
was incorrectly marked in the header as having units of
pixels per inch (hex 01), where it should have indicated
no units (hex 00).  Now, such an image is correctly marked
as having no density units, which is a one-bit difference in
the file but unfortunately may bump the size of the git repo
and the ostree differences.

https://phabricator.endlessm.com/T15193