report-product-apps: add Southeast Asia locales
The Southeast Asia image supports four locales:
- English (for use in India)
- Indonesian
- Thai
- Vietnamese

The image name is "sea", and I display app details in English
for that locale, but then I added additional locales for
id, th, and vi, which include app detail translations to
those languages.

Note that with the current implementation, only the English
version of the encyclopedia is listed in the results.