Exclude translations for localized app IDs
Any "com.endlessm" app ID that includes a locale suffix
is expected to already have localized data in the CMS,
so we don't want to bother translators in Transifex
with lots of content that is specific to a single language.

This also avoids a silly step of translating this content
from its original language into English, only so that it
can be retranslated into the original language in Transifex.

https://phabricator.endlessm.com/T16288