Merge translation info into content.json
Add a new data file, bundle/translations.json, that has information that
will be used to add <translation> tags to the AppData XML files. The
translation_type and translation_id keys will be merged into
content.json and pulled out again at bundling time. Only a few of the
apps currently have correct settings, but the keys are needed for the
AppData work to proceed.

https://phabricator.endlessm.com/T16227