eosshellcontent.py: Fix generating screenshots for the appdata files
The screenshots were not being correctly generated as several different
images were being set per screenshot due to a confusion of how the
screenshots are arranged in the metadata and in the appdata files.

The metadata arranges images per language but these changes group the
languages per image instead, and produce screenshots for each one.

https://phabricator.endlessm.com/T15856