eos-content-merge: Ignore localized icons
Upstream GNOME apps have started to provide localization
of the icon specification in desktop files as part of
the migration from intltool to gettext, even though
in practice they specify the same icon for all locales.
Let's filter these localized icons out so that the
Endless provided icon is used for all locales.

https://phabricator.endlessm.com/T17971