report-product-apps: include gnome repo
Although we don't have CMS data, let's at least include
the apps as available.

We have to omit the spurious freedesktop GlxInfo which does not
show up in our app center, and all the apps that we blacklist
in gnome-software (due to duplication with core/installed apps
and/or broken apps).