unzip_content: fix extra categories for knowledge apps
This was lost when we changed the CMS to use full localized app IDs.

If no extra categories for the full app ID, see if there are extra
categories listed for the generic ID without the locale extension.

While we could have explicitly listed all the localized app IDs
in extra_categories.py, that would be fragile to the expectation
that we will continue to add new localized versions of some
of these apps.

https://phabricator.endlessm.com/T18549