build: Generate json.h files in srcdir
po/Makefile is generated assuming all POTFILES are in the source
directory, but the json.h strings files were being created in the build
directory. This was causing distcheck to fail on automake 1.15 where an
additional build subdirectory is used to catch exactly these types of
errors. See [1] for details.

In addition to generating the files in the source directory, distribute
them since there's no reason to regenerate them at build time. If the
source files change, then make will regenerate them on an incremental
build. This helps ensure that the po files can be generated without
previously running make.

1. https://debbugs.gnu.org/cgi/bugreport.cgi?bug=18286

https://phabricator.endlessm.com/T18550