build: Remove json.h dependency on extract-content-strings
With the change to generate the json.h files in the source directory,
they can't depend on anything that only gets generated at build time.
That would cause the files to be uselessly regenerated during the build,
which breaks distcheck since it uses a read-only source directory.

This is a minor annoyance since you'd like changes to the extractor to
automatically regenerate the needed files. In that case,
maintainer-clean can be used to remove the json.h files and start clean.

https://phabricator.endlessm.com/T18550