report-product-apps: Optionally include Flathub
All the apps that are available in gnome-apps are now available
in flathub, so use flathub rather than gnome-apps.  However, this
list is a bit long and does not filter out the many runtimes nor
those apps that we blacklist due to duplications or problems,
so we don't want to include them by default.  We add an optional
"--flathub" option to enable this list.