Fix com.endlessnetwork.MidnightmareTeddy app ID
App IDs are case-sensitive. In a list of these app IDs, this one was
claimed to be lowercase; but what was actually submitted to Flathub
<https://github.com/flathub/com.endlessnetwork.MidnightmareTeddy> is in
CamelCase.

The other two which are already on Flathub – Aqueducts and Dragon's
Apprentice – are in the correct case.

https://phabricator.endlessm.com/T25938
(cherry picked from commit 2381d993ff74d30ee55d50262e3b984b4b62b373)