icon grid: Remove Translation
I've seen screenshots that show that this app used to look a lot better,
but at some point it has stopped working well. Unless/until we spend
time improving it, drop it from the default desktop. (It's not generally
installed anyway. Perhaps we could just replace it with a web link for
Google Translate?)