build: Make json.h files during normal build
Use automake's `dist_noinst_DATA` primary so that the json.h files get
built with make's default targets and not just `dist`. As before the
files are still distributed so that they wouldn't need to be regenerated
in OBS when a tarball is in use. This allows our transifex tooling to
run a normal checkout prepared with `make` instead of the full `make
distcheck`.

https://phabricator.endlessm.com/T27529