eos-save-icon-grid: Rewrite script to read new-style settings
Previously, this script dumped the old `icon-grid-layout` GSettings key
into a JSON file, so that the JSON file could be put into the image
configuration (using the `63-icon-grid` hook in eos-image-builder, which
loads data from JSON files stored in
`endless-image-config.git/data/icon-grid`).

The `icon-grid-layout` key has been removed with the reworking and
upstreaming of the customisable icon grids to gnome-shell. The new icon
grid configuration is stored in `app-picker-layout` and relocatable
schemas below `/org/gnome/desktop/app-folders/folders/`.

Rework this script to load its data from the new icon grid configuration
instead, and convert it to the existing JSON format to be loaded by
eos-image-builder in the same way as before.

The conversion from the new icon grid configuration format is lossy, so
this is not the best long-term solution. In the future, it would be
better to ship the icon grid configuration as dconf overrides. However,
that’s a non-trivial amount of work and requires changes to the image
builder which are not sensible at this time in the release cycle.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T31616