icon grid: Remove per-architecture blacklist
The keys in arch-blacklist.json are the autotools $(host_cpu) variable.
The only key in that file is "arm". But nowadays we only support 64-bit
ARM, for which $(host_cpu) is "aarch64". So this filter now does
nothing.

It doesn't matter, anyway, because the app grid code in Endless OS 3.9
and later just hides empty folders.

Remove the blacklist logic, which also removes some problematic
terminology from our distro.