icon grid: Remove Chromium & Nautilus
These are on the dash by default, configured via an override of the
org.gnome.shell favorite-apps GSetting in eos-theme.

In GNOME 40 and newer, an app cannot be both in the grid and on the dash
at once, and the dash takes precedence.

The code in eos-desktop-extension which imports these JSON files doesn't
know about this mutual-exclusion, so it reserves space for these apps on
the grid even though they don't appear there, leading to a weird space
at the end of the first page.

https://phabricator.endlessm.com/T34646