icon grid: Fix Kolibri channel desktop file IDs
The Kolibri Flatpak contains a plugin that exports a .desktop file for
each installed channel. Previously, these files had names of the form:

    org.learningequality.Kolibri.Channel.${CHANNEL_ID}.desktop

However, if the CHANNEL_ID begins with a digit, the resulting name
(minus .desktop) is not a valid D-Bus name, since components of bus
names can't begin with a digit. So [the scheme was changed][0]
to:

    org.learningequality.Kolibri.channel_${CHANNEL_ID}.desktop

i.e. lower-casing the C, and replacing the penultimate dot with an
underscore.

Adjust all the grids accordingly.

[0]: https://github.com/endlessm/kolibri-app-desktop-xdg-plugin/commit/8e0b21d2888f04f168452eadf021e2985cc585fc

https://phabricator.endlessm.com/T34646